09: Split get_param into get_param and get_address
Rather than passing in a `write` param that tells the function that what
we want is an address that we're going to write to, let's just define a
separate function for that use case. That way we can have different
return types and also panic whenever we expect to fetch an address to
write to but that parameter is in Immediate mode (which is not allowed
as per the instructions from day 05).